// frontend/src/client/types.gen.ts
export type UserPublic = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    full_name?: (string | null);
    id: string;
    role?: (UserRole | null);
};

export type UserRegister = {
    email: string;
    password: string;
    full_name?: (string | null);
};

export type UserRole = 'doctor' | 'patient';